Split E2E teams tests into dedicated spec and page object
Extracted team CRUD tests from 04-groups.spec.js into 04b-teams.spec.js
with a new TeamsPage.js page object model, matching the R8 pages/teams/
directory reorganization. Groups/Teams switching tests remain in
04-groups.spec.js since they span both domains.
